Refactor: Break down Login page into components
Refactor the Login page into smaller, reusable components to improve maintainability and readability.
